chore: move youtube helper to music, add twitch permissions, remove scaffold (#442)
- Move youtube() and its googleapiclient import from utils/helpers.py
  to modules/music/playback.py, its sole caller. Removes the heavy
  optional dependency from every import of utils.helpers.
- Add twitch entry to REQUIRED_PERMISSIONS in utils/permissions.py
  (send_messages, embed_links), matching the permissions the module
  already uses in practice.
- Delete three unused Vite/Vue scaffold files: HelloWorld.vue, vue.svg,
  vite.svg. Never referenced in the app.
